[Oztechan/CCC#1842] Create FormatUtil in :client:core:viewmodel and move relevant methods from client
